chore: move constraints to a separate file
